Move instance to own file and fix imports
