+++ types/register.d.ts
@@ -1,7 +1,7 @@
 import { RouteShorthandOptions } from './route'
 import { HTTPMethods, RawServerBase, RawServerDefault, RawRequestBase, RawRequestDefault, RawReplyBase, RawReplyDefault } from './utils'
 import { FastifyPlugin } from './plugin';
-import { FastifyInstance } from '../fastify';
+import { FastifyInstance } from './instance';
 
 export type FastifyRegister<
   RawServer extends RawServerBase = RawServerDefault,
